fix clean messages

+++ src/commands/Diversos.py
@@ -133,7 +133,15 @@ class Diversos(commands.Cog):
     @commands.command(help="Limpar chat")
     async def clean(self, ctx, amount):
 
-        async for message in ctx.channel.history(limit = int(amount) + 1):
+        if self.bot.eh_plebe(ctx.author):
+            return await ctx.send("Seu pau é infelizmente muito pequeno para utilizar este comando")
+            
+        amount = int(amount)
+
+        if amount > 100:
+            return await ctx.send('menos de 100 mensagens por favor...')
+
+        async for message in ctx.channel.history(limit = amount + 1):
             await message.delete()
 
         return await ctx.send(f'foram deletadas {amount} mensagens')
